[feature] 1.7

+++ app/src/main/java/com/heyanle/closure/net/model/WebsiteUser.kt
@@ -5,7 +5,7 @@ package com.heyanle.closure.net.model
  * https://github.com/heyanLE
  */
 data class WebsiteUser(
-    var bindQQ: Int,
+    var bindQQ: Long,
     var email: String,
     var isAdmin: Boolean,
     var password: String,
